+++ packages/server/dataloader/customLoaderMakers.ts
@@ -3,10 +3,10 @@ import {Selectable, SqlBool, sql} from 'kysely'
 import {PARABOL_AI_USER_ID} from '../../client/utils/constants'
 import MeetingTemplate from '../database/types/MeetingTemplate'
 import getFileStoreManager from '../fileStorage/getFileStoreManager'
+import isValid from '../graphql/isValid'
 import {ReactableEnum} from '../graphql/public/resolverTypes'
 import {SAMLSource} from '../graphql/public/types/SAML'
 import getKysely from '../postgres/getKysely'
-import {TeamMeetingTemplate} from '../postgres/pg.d'
 import {IGetLatestTaskEstimatesQueryResult} from '../postgres/queries/generated/getLatestTaskEstimatesQuery'
 import getGitHubAuthByUserIdTeamId, {
   GitHubAuth
@@ -29,12 +29,14 @@ import {
 } from '../postgres/select'
 import {Insight, MeetingSettings, OrganizationUser, Task, Team} from '../postgres/types'
 import {AnyMeeting, MeetingTypeEnum} from '../postgres/types/Meeting'
+import {TeamMeetingTemplate} from '../postgres/types/pg'
 import {Logger} from '../utils/Logger'
 import getRedis from '../utils/getRedis'
 import isUserVerified from '../utils/isUserVerified'
 import NullableDataLoader from './NullableDataLoader'
 import RootDataLoader, {RegisterDependsOn} from './RootDataLoader'
 import normalizeArrayResults from './normalizeArrayResults'
+import normalizeResults from './normalizeResults'
 export interface MeetingSettingsKey {
   teamId: string
   meetingType: MeetingTypeEnum
@@ -154,6 +156,11 @@ export const userTasks = (parent: RootDataLoader, dependsOn: RegisterDependsOn) 
             filterQuery,
             includeUnassigned
           } = key
+          if (teamIds.length === 0)
+            return {
+              key: serializeUserTasksKey(key),
+              data: []
+            }
           const hasUserIds = userIds?.length > 0
           const hasStatusFilters = statusFilters ? statusFilters.length > 0 : false
           const teamTasks = await selectTasks()
@@ -579,16 +586,28 @@ export const lastMeetingByMeetingSeriesId = (
   dependsOn('newMeetings')
   return new DataLoader<number, AnyMeeting | null, string>(
     async (keys) => {
-      return await Promise.all(
-        keys.map(async (key) => {
-          const latestMeeting = await selectNewMeetings()
-            .where('meetingSeriesId', '=', key)
-            .orderBy('createdAt desc')
-            .limit(1)
-            .executeTakeFirst()
-          return latestMeeting || null
-        })
-      )
+      const meetingIdRes = await getKysely()
+        .with('LastMeetings', (qc) =>
+          qc
+            .selectFrom('NewMeeting')
+            .select([
+              'id',
+              'meetingSeriesId',
+              'createdAt',
+              sql`ROW_NUMBER() OVER (PARTITION BY "meetingSeriesId" ORDER BY "createdAt" DESC)`.as(
+                'rn'
+              )
+            ])
+            .where('meetingSeriesId', 'in', keys)
+        )
+        .selectFrom('LastMeetings')
+        .select('id')
+        .where('rn', '=', 1)
+        .execute()
+
+      const meetingIds = meetingIdRes.map(({id}) => id)
+      const meetings = (await parent.get('newMeetings').loadMany(meetingIds)).filter(isValid)
+      return normalizeResults(keys, meetings, 'meetingSeriesId')
     },
     {
       ...parent.dataLoaderOptions