+++ packages/server/postgres/getPg.ts
@@ -1,7 +1,17 @@
-import {Pool} from 'pg'
+import type {Pool as PGPool} from 'pg'
 import sleep from '../../client/utils/sleep'
 import getPgConfig from './getPgConfig'
+/*
+Force a native `require` so dd-trace-js can monkeypatch the require statement.
+In development, the require statement requires `./pg.ts` since require resolves packages by first looking in the same dir
+In production, the require statement will resolve to the node_modules found in /dist
+*/
 
+// when used outside of webpack, e.g. kysely.config.js, go vanilla
+declare let __non_webpack_require__: typeof require
+const pg =
+  typeof __non_webpack_require__ === 'undefined' ? require('pg') : __non_webpack_require__('pg')
+const {Pool} = pg
 const config = getPgConfig()
 
 const graceFullyReconnect = async () => {
@@ -21,10 +31,10 @@ const graceFullyReconnect = async () => {
   }
 }
 
-let pool: Pool | undefined
+let pool: PGPool | undefined
 const getPg = (schema?: string) => {
   if (!pool) {
-    pool = new Pool(config)
+    pool = new Pool(config) as PGPool
     pool.on('error', graceFullyReconnect)
     if (schema) {
       pool.on('connect', (client) => {